Clean up and move Metric class into own file.
